Extract AbstractTestHiveViews
